+++ js/config.js
@@ -1,0 +1,40 @@
+export const MOCK_COMMENTS = [
+  "Всё отлично!",
+  "В целом всё неплохо. Но не всё.",
+  "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
+  "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
+  "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
+  "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!",
+];
+
+export const MOCK_DESCRIPTIONS = [
+  "Закат над горным озером с отражением розовых облаков в спокойной воде",
+  "Старинная европейская улочка с брусчаткой и цветущими балконами",
+  "Черно-белый портрет пожилого мужчины с выразительными морщинами у глаз",
+  "Макросъемка капель росы на лепестках красной розы ранним утром",
+  "Городской пейзаж с небоскребами в золотых лучах заходящего солнца",
+];
+
+export const MOCK_NAMES = [
+  "Алексей",
+  "Мария",
+  "Дмитрий",
+  "Анна",
+  "Сергей",
+  "Елена",
+  "Владимир",
+  "Ольга",
+  "Андрей",
+  "Наталья",
+  "Павел",
+  "Ирина",
+  "Николай",
+  "Татьяна",
+  "Михаил",
+];
+
+export const CONFIG = {
+  names: MOCK_NAMES,
+  descriptions: MOCK_DESCRIPTIONS,
+  comments: MOCK_COMMENTS,
+};
